refactor(editor): extract handleSave into useDocumentSave (Spec #6 crown jewel)
Moves the document SAVE path — handleSave plus its two save-only helpers
(injectReplyRangeMarkers, injectTCReplyRangeMarkers) — out of the 12k-line
DocxEditor god-component into a useDocumentSave hook (~267 lines).

This is the most load-bearing IO code (a bug is data loss), so it's a VERBATIM
move with the IDENTICAL dep array [onSave, emitError, emitEvent, comments],
preserving the invariants by construction:
 - comments stays a by-value dep (a stale snapshot drops new comments/replies);
 - the refs (agent/paged/footnote/endnote/props) are passed by reference and
   read as .current at save time — never snapshotted;
 - the mutation order is unchanged (content -> comments -> footnotes -> endnotes
   -> properties -> reply markers -> selective options -> toBuffer -> selective
   clearTrackedChanges -> onSave -> emitEvent).

The callers (download/autosave/make-copy/email/export-as, imperative save) stay
in the component and use the hook's returned handleSave.

Verified: typecheck + eslint (rules-of-hooks + dep array) + 439 react tests +
round-trip gate pristine (47 fixtures) + 47 save e2e — handleSave
characterization (tracked-change save->reload), footnote & endnote save->reload,
demo-docx round-trip save — 0 fail.
