+++ docx-editor/.changeset/use-document-save.md
@@ -1,0 +1,5 @@
+---
+'@casualoffice/docs': patch
+---
+
+Internal: extract the document save path (`handleSave`) out of the DocxEditor component into a `useDocumentSave` hook. Verbatim move, no behaviour change.
